requirements.txt
ruff
psutil
gputil
pre-commit
GitPython
